test: move network integration tests into testing
