+++ testing/src/network/sync.rs
@@ -13,10 +13,12 @@
 
 // You should have received a copy of the GNU General Public License
 // along with the snarkOS library. If not, see <https://www.gnu.org/licenses/>.
-mod common;
-use common::start_node;
 
-use snarkos_testing::consensus::{BLOCK_1, BLOCK_2};
+use crate::{
+    consensus::{BLOCK_1, BLOCK_2, TRANSACTION_1},
+    network::start_node,
+};
+
 use snarkvm_objects::block::Block;
 
 use std::time::Duration;
@@ -78,4 +80,41 @@ async fn simple_block_sync() {
             .read()
             .block_hash_exists(&block_struct_2.header.get_hash())
     );
+}
+
+#[tokio::test]
+async fn simple_transaction_sync() {
+    use snarkos_consensus::memory_pool::Entry;
+    use snarkvm_dpc::instantiated::Tx;
+    use snarkvm_utilities::bytes::FromBytes;
+
+    let node_alice = start_node(vec![]).await;
+    let alice_address = node_alice.local_address().unwrap();
+
+    // insert transaction into node_alice
+    let mut memory_pool = node_alice.environment.memory_pool().lock();
+    let storage = node_alice.environment.storage().read();
+
+    let transaction = Tx::read(&TRANSACTION_1[..]).unwrap();
+    let size = TRANSACTION_1.len();
+    let entry = Entry {
+        size_in_bytes: size,
+        transaction: transaction.clone(),
+    };
+
+    memory_pool.insert(&storage, entry.clone()).unwrap().unwrap();
+
+    // drop the locks to avoid deadlocks
+    drop(memory_pool);
+    drop(storage);
+
+    let node_bob = start_node(vec![alice_address.to_string()]).await;
+
+    // T 0-2s: not much happens
+    // T 2s: first sync occures, a peer isn't yet connected to sync with
+    // T 4s: second sync occures, this time a peer is selected for the block sync
+    sleep(Duration::new(5, 0)).await;
+
+    // check transaction is present in bob's memory pool
+    assert!(node_bob.environment.memory_pool().lock().contains(&entry));
 }
